// src/app/services/user/model/user-model.ts
export interface UserModel{
  id: string;
  createdAt: string;
  Name: string;
  Avatar: string;
  City: string;
  PostCode: string;
  JobTitle: string;
  PhoneNumber: string;
  Email: string;
  Password: string;
  Latitude: string;
  Longitude: string;
  BackgrounImage: string;
  Vehicle: string;
  BirthDate: string;
  Country: string;

}